# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ActivitySummary:
    properties:
      action_type:
        type: string
      created_at:
        type: string
      id:
        type: integer
      preview:
        type: string
      target_id:
        type: integer
      target_type:
        type: string
    type: object
...
  dto.DiaryEntry:
    properties:
      game:
        $ref: '#/definitions/dto.GameSummary'
      logged_at:
        type: string
      status:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      field:
        type: string
    type: object
  dto.FollowRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  dto.FollowResponse:
    properties:
      is_following:
        type: boolean
    type: object
...
  dto.GameSummary:
    properties:
      avg_rating:
        type: number
      id:
        type: integer
      is_free:
        type: boolean
      poster:
        type: string
      price:
        type: number
      release_date:
        type: string
      review_count:
        type: integer
      steam_id:
        type: integer
      title:
        type: string
    type: object
...
  dto.PaginatedResponse-array_dto_DiaryEntry:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DiaryEntry'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
  dto.PaginatedResponse-array_dto_GameSummary:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.GameSummary'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
...
  dto.PaginatedResponse-array_dto_UserSummary:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserSummary'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
  dto.PaginationDTO:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  dto.ReviewSummary:
    properties:
      comment_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      game:
        $ref: '#/definitions/dto.GameSummary'
      id:
        type: integer
      img:
        type: string
      is_spoiler:
        type: boolean
      like_count:
        type: integer
      recommend:
        type: string
      title:
        type: string
    type: object
...
  dto.SuccessResponse-dto_FollowResponse:
    properties:
      data:
        $ref: '#/definitions/dto.FollowResponse'
    type: object
...
  dto.SuccessResponse-dto_UserProfileResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserProfileResponse'
    type: object
  dto.SuccessResponse-dto_UserStatsResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserStatsResponse'
    type: object
...
  dto.UserProfileResponse:
    properties:
      avatar_url:
        type: string
      average_rating:
        type: number
      backlog_games:
        items:
          $ref: '#/definitions/dto.GameSummary'
        type: array
      bio:
        type: string
      created_at:
        type: string
      diary:
        items:
          $ref: '#/definitions/dto.DiaryEntry'
        type: array
      email:
        type: string
      follower_count:
        type: integer
      following_count:
        type: integer
      game_logs_count:
        type: integer
      id:
        type: integer
      list_count:
        type: integer
      popular_reviews:
        items:
          $ref: '#/definitions/dto.ReviewSummary'
        type: array
      recent_activity:
        items:
          $ref: '#/definitions/dto.ActivitySummary'
        type: array
      recent_reviews:
        items:
          $ref: '#/definitions/dto.ReviewSummary'
        type: array
      review_count:
        type: integer
      role:
        $ref: '#/definitions/models.UserRole'
      steam_id:
        type: string
      username:
        type: string
    type: object
...
  dto.UserStatsResponse:
    properties:
      average_rating:
        type: number
      genre_distribution:
        additionalProperties:
          type: integer
        type: object
      rating_distribution:
        additionalProperties:
          type: integer
        description: 1 to 5 stars
        type: object
      total_played:
        type: integer
      total_reviews:
        type: integer
    type: object
  dto.UserSummary:
    properties:
      avatar:
        type: string
      bio:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  models.UserRole:
    enum:
    - user
    - influencer
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleInfluencer
    - RoleAdmin
host: localhost:8080
info:
  contact: {}
  description: Hệ thống API cho Gaming Box.
  title: Gaming Box API
  version: "1.0"
paths:
...
  /users/diary:
    get:
      description: Lấy danh sách nhật ký chơi game của người dùng
      parameters:
      - description: ID người dùng
        in: query
        name: id
        type: integer
      - default: 1
        description: Trang
        in: query
        name: page
        type: integer
      - default: 20
        description: Giới hạn
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_DiaryEntry'
      summary: Xem Nhật ký chơi game
      tags:
      - Users
  /users/follow:
    post:
      consumes:
      - application/json
      description: Thực hiện follow hoặc unfollow một người dùng khác (Cần đăng nhập)
      parameters:
      - description: Thông tin follow
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow / Unfollow Người dùng khác
      tags:
      - Users
  /users/me:
    get:
      description: Lấy thông tin profile đầy đủ của người dùng (Cần đăng nhập)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy Profile Người dùng
      tags:
      - Users
  /users/me/followers:
    get:
      description: Lấy danh sách những người dùng đang follow current user
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số lượng hiển thị trên mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_UserSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách người theo dõi
      tags:
      - Users
  /users/me/following:
    get:
      description: Lấy danh sách những người dùng mà current user đang follow
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số lượng hiển thị trên mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_UserSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách đang theo dõi
      tags:
      - Users
  /users/profile:
    get:
      description: Lấy thông tin profile đầy đủ của một người dùng dựa theo ID
      parameters:
      - description: ID của Người dùng
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Xem Profile người dùng khác
      tags:
      - Users
  /users/stats:
    get:
      description: Xem thống kê về game đã chơi, đánh giá trung bình...
      parameters:
      - description: ID người dùng (Nếu không có sẽ lấy me)
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserStatsResponse'
      summary: Lấy thống kê người dùng
      tags:
      - Users
  /users/watchlist:
    get:
      description: Lấy danh sách game muốn chơi của người dùng
      parameters:
      - description: ID người dùng
        in: query
        name: id
        type: integer
      - default: 1
        description: Trang
        in: query
        name: page
        type: integer
      - default: 20
        description: Giới hạn
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameSummary'
      summary: Xem Watchlist (Backlog)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"